requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blis==0.7.11
Brotli==1.0.9
bs4==0.0.2
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
cloudpickle==3.1.0
compressed-tensors==0.6.0
confection==0.1.5
dataclasses-json==0.6.7
datasets==3.0.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
einops==0.8.0
exceptiongroup==1.2.2
fastapi==0.112.4
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.6.1
gguf==0.10.0
gmpy2==2.1.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.23.5
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.4.5
interegular==0.3.3
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonargparse==4.32.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.28
langchain-text-splitters==0.2.2
langcodes==3.4.0
langsmith==0.1.88
language_data==1.2.0
lark==1.2.2
Levenshtein==0.25.1
lightning==2.3.3
lightning-utilities==0.11.6
llvmlite==0.41.1
lm-format-enforcer==0.10.6
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.3
mdurl==0.1.2
mistral_common==1.4.4
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
numba==0.58.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
openai==1.52.0
opencv-python-headless==4.10.0.84
orjson==3.10.6
outlines==0.0.46
packaging==24.1
pandas==2.0.3
partial-json-parser==0.2.1.1.post4
pathlib_abc==0.1.1
pathy==0.11.0
pillow==10.4.0
pip==24.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
pooch==1.7.0
prometheus_client==0.21.0
prometheus-fastapi-instrumentator==7.0.0
protobuf==5.28.2
psutil==6.1.0
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==17.0.0
pycountry==24.6.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.2.0
rank-bm25==0.2.2
rapidfuzz==3.9.4
ray==2.10.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.0.1
sentencepiece==0.2.0
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
starlette==0.38.6
sympy==1.12
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.1
torch==2.4.0
torchmetrics==1.4.1
tqdm==4.66.4
transformers==4.45.2
triton==3.0.0
typer==0.4.2
typing_extensions==4.11.0
typing-inspect==0.9.0
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.32.0
uvloop==0.21.0
vllm==0.6.3.post1
wasabi==0.10.1
watchfiles==0.24.0
weasel==0.4.1
websockets==13.1
wheel==0.43.0
wrapt==1.16.0
xformers==0.0.27.post2
xxhash==3.5.0
yarl==1.9.4
zipp==3.20.2
